feat: ( #1692 )  convert  mine integration page to kotlin

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/wallet/integration/mine/MineIntegrationContract.java
@@ -23,7 +23,7 @@ public interface MineIntegrationContract {
          *
          * @param balance current user's balance
          */
-        void updateBalance(double balance);
+        void updateBalance(long balance);
 
         /**
          * handle request loading
